+++ macros/spells/darkness.js
@@ -1,27 +1,36 @@
-if (args[0].tag === "OnUse" && ["preTargeting"].includes(args[0].macroPass)) {
+const functionArgs = (typeof args !== 'undefined') ? args : undefined;
+
+if (functionArgs && args[0].tag === "OnUse" && ["preTargeting"].includes(args[0].macroPass)) {
   args[0].workflow.item.system['target']['type'] = "self";
   args[0].workflow.item.system.range = { value: null, units: "self", long: null };
   return;
 }
 
-const lastArg = args[args.length - 1];
-const tokenOrActor = await fromUuid(lastArg.actorUuid);
-const targetActor = tokenOrActor.actor ? tokenOrActor.actor : tokenOrActor;
-
-if (args[0] === "on") {
+async function placeTemplate({ origin, targetActor, targetToken } = {}) {
   Hooks.once("createMeasuredTemplate", async (template) => {
-    let radius = canvas.grid.size * (template.data.distance / canvas.grid.grid.options.dimensions.distance);
+
+    const v12 = foundry.utils.isNewerVersion(game.version, 12);
+    const data = v12 ? template : template.data;
+    const grid = v12 ? canvas.grid : canvas.grid.grid.options.dimensions;
+    console.warn(template)
+    let radius = canvas.grid.size * (data.distance / grid.distance);
     const darknessSpellParams = {
       radius,
-      x: template.data.x,
-      y: template.data.y,
-      distance: template.data.distance,
+      x: data.x,
+      y: data.y,
+      distance: data.distance,
       targetActorId: targetActor.id,
     };
-    await DAE.setFlag(targetActor, "darknessSpell", darknessSpellParams);
+    // await DAE.setFlag(targetActor, "darknessSpell", darknessSpellParams);
+    await targetActor.update({
+      "flags.world.darknessSpell": {
+        active: true,
+        params: darknessSpellParams,
+      },
+    });
     canvas.scene.deleteEmbeddedDocuments("MeasuredTemplate", [template.id]);
     // console.warn("Darkness Spell", darknessSpellParams);
-    await DDBImporter.lib.DDBMacros.executeDDBMacroAsGM("gm", "darkness", { origin: lastArg.origin, effect: lastArg.origin }, { args: ["on", darknessSpellParams] });
+    await DDBImporter.lib.DDBMacros.executeDDBMacroAsGM("gm", "darkness", { origin: origin, effect: origin }, { args: ["on", darknessSpellParams] });
   });
 
   const measureTemplateData = {
@@ -45,10 +54,55 @@ if (args[0] === "on") {
   const measureTemplate = new game.dnd5e.canvas.AbilityTemplate(doc);
   measureTemplate.actorSheet = targetActor.sheet;
   measureTemplate.drawPreview();
-} else if (args[0] === "off") {
-  const params = await DAE.getFlag(targetActor, "darknessSpell");
-  // const gmMacro = game.macros.find((m) => m.name === gmMacroName);
-  // gmMacro.execute({ actor, token, args: ["off", params] })
-  DDBImporter.lib.DDBMacros.executeDDBMacroAsGM("gm", "darkness", { actor: actor._id, token: token._id }, { args: ["off", params] });
-  await DAE.unsetFlag(targetActor, "darknessSpell");
+}
+
+const isSimpleDDBMacro = scope && foundry.utils.getProperty(scope, "flags.ddb-importer.ddbMacroFunction");
+
+if (isSimpleDDBMacro) {
+
+
+  if (!actor || !item) {
+    console.warn("no actor or item selected", actor, item);
+    return;
+  }
+}
+
+console.warn("SimpleMacro", isSimpleDDBMacro);
+
+const lastArg = isSimpleDDBMacro
+  ? {}
+  : args[args.length - 1];
+const tokenOrActor = isSimpleDDBMacro
+  ? token ?? actor
+  : await fromUuid(lastArg.actorUuid);
+const targetActor = isSimpleDDBMacro
+  ? actor
+  : tokenOrActor.actor ? tokenOrActor.actor : tokenOrActor;
+
+const origin = isSimpleDDBMacro
+  ? item
+  : await lastArg.origin;
+
+const flagData = foundry.utils.getProperty(actor, "flags.world.darknessSpell");
+
+const isOn = isSimpleDDBMacro
+  ? (!flagData?.active ?? true)
+  : args[0] === "on";
+
+const isOff = isSimpleDDBMacro
+  ? (flagData?.active ?? false)
+  : args[0] === "off";
+
+if (isOn) {
+  await placeTemplate({ origin, targetActor, targetToken: tokenOrActor });
+} else if (isOff) {
+  // const params = await DAE.getFlag(targetActor, "darknessSpell");
+  DDBImporter.lib.DDBMacros.executeDDBMacroAsGM("gm", "darkness", { actor: actor._id, token: token._id }, { args: ["off", flagData.params] });
+  await targetActor.update({
+    "flags.world.darknessSpell": {
+      active: false,
+      params: null,
+    },
+  });
+  // await DAE.unsetFlag(targetActor, "darknessSpell");
 }